# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CategoryResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  dto.CommentResponse:
    properties:
      blogId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      parentId:
        type: string
      replies:
        items:
          $ref: '#/definitions/dto.CommentResponse'
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/dto.UserBriefResponse'
      userId:
        type: string
    type: object
...
  dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      slug:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    - slug
    type: object
  dto.CreateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
      parentId:
        type: string
    required:
    - content
    type: object
...
  dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      slug:
        maxLength: 100
        minLength: 1
        type: string
    type: object
  dto.UpdateCommentRequest:
    properties:
      content:
        minLength: 1
        type: string
    required:
    - content
    type: object
...
  dto.UserBriefResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Clean Architecture Go API with GORM, Gin, PostgreSQL, Redis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /api/v1/blogs/{blogId}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for a blog post
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get comments for a blog
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a comment on a blog post
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a comment
      tags:
      - Comments
...
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: Get all categories
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new blog category
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a category
      tags:
      - Categories
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Get a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a category
      tags:
      - Categories
  /api/v1/comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment (owner only)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a comment (owner only)
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a comment
      tags:
      - Comments
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"